Improvementsand new install environment scripts

+++ pages/Dagelijks.py
@@ -3,9 +3,15 @@ import logging
 import json
 import os
 import shutil
-# Initialize logging to log to a rolling file in log directory with a max size of 10 MB and a backup count of 7 files
-logging.basicConfig(filename='log/dagelijks.log', level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+from datetime import datetime
+# Initialize logging to log to a rolling file in log directory with a max size of 10 MB and a backup count of 7 files filename timestamped with current date
+# Create the log directory if it doesn't exist
+if not os.path.exists('log'):
+    os.makedirs('log')
+   
+logging.basicConfig(filename=f"log/dagelijks{datetime.now().strftime('%Y%m%d')}.log", level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
+plan_file = "plan/2025M01_plan.json"
 
 # a method to unzip a file
 def unzip_file(zip_file, target_folder):
@@ -69,63 +75,72 @@ def copy_files_from_plan(plan_file, target_folder):
                 logging.info(f"Copying {source} to {target}")
                 st.info(f"Copying {source} to {target}")
         
-        shutil.copy(source, target)
+        # if source file does not exist display a warning and skip the copy
+        if not os.path.exists(source):
+            logging.error(f"Source file {source} does not exist. Skipping copy.")
+            st.error(f"❌ Source file {source} does not exist. Skipping copy.")
+            continue
+        else:
+            # Copy the file from source to target
+            shutil.copy(source, target)
             
         # Update the size and md5 in the plan
         file_info["size"] = os.path.getsize(target)
         file_info["md5"] = calculate_md5(target)
         logging.info(f"Copied {source} to {target}, size: {file_info['size']}, md5: {file_info['md5']}")
 
-
-monthly_period = st.selectbox("Select the period:", ["202501", "202502", "202503", "202504", "202505", "202506", "202507", "202508", "202509", "202510", "202511", "202512"])
-quarterly_period = st.selectbox("Select the quarterly period:", ["2025Q01", "2025Q02", "2025Q03", "2025Q04"])
-
 st.title("Dagelijks")
 
-files = st.text_area("Enter the list of files to copy (e.g., file1.csv):", "C:\\Repos\\Data\\Curves\\Book_3.xlsx\nC:\\Repos\\Data\\Curves\\Book_2.xlsx")
-lines = files.split("\n")
+col1, col2 = st.columns(2)
 
-for line in lines:
-    if line.strip():  # Check if the line is not empty
-        logging.info(f"Processing file: {line.strip()}")
-
-        # If file exists write to log info and display checked otherwise write to log error and display error message
-        try:
-            with open(line.strip(), 'r') as f:
-                logging.info(f"File {line.strip()} exists.")
-                st.info(f"✅ File {line.strip()} exists.")
-        except FileNotFoundError:
-            logging.error(f"File {line.strip()} does not exist.")
-            st.error(f"❌ File {line.strip()} does not exist.")
-            continue
-        except Exception as e:
-            logging.error(f"Error processing file {line.strip()}: {e}")
-            st.error(f"❌ Error processing file {line.strip()}: {e}")
-            continue    
-
-if st.button("Process File", key=line.strip()):
-    # Copy Files to Target
-    logging.info(f"File {line.strip()} processed successfully.")
-    st.success(f"✅ File {line.strip()} processed successfully.")
+monthly_period = col1.selectbox("Select the period:", ["202501", "202502", "202503", "202504", "202505", "202506", "202507", "202508", "202509", "202510", "202511", "202512"])
+quarterly_period = col2.selectbox("Select the quarterly period:", ["2025Q01", "2025Q02", "2025Q03", "2025Q04"])
 
 
+# Read the plan file to a string with indented 4 space JSON
+with open(plan_file, 'r') as f:
+    plan_content = json.dumps(json.load(f), indent=4)
+    st.code(plan_content, language='json', height=300)
 
-    
+
+if st.button("Verify Source Files"):
+    plan = []
+    # Read the plan file and load the JSON data
+    with open(plan_file, 'r') as f:
+        plan = json.load(f)
+    for file_info in plan:
+        source = file_info["source"]
+        
+        # verify the source file exists
+        if os.path.exists(source): 
+            logging.info(f"Source file {source} exists.")
+            st.success(f"✅ Source file {source} exists.")
+        else:
+            logging.error(f"Source file {source} does not exist.")
+            st.error(f"❌ Source file {source} does not exist.")
 
 
-if st.button("Export File List"):
-    # Generate file list and export to JSON
-    source_files = ["C:\\Repos\\Data\\Curves\\Book_3.xlsx", "C:\\Repos\\Data\\Curves\\Book_2.xlsx", "C:\\Repos\\Data\\Curves\\Book_1.xlsx"]
-    target_folder = "C:\\Repos\\Data\\Curves\\Target"
-    file_list = generate_file_list(source_files, target_folder)
-    output_file = "plan/2025M01_plan.json"
-    export_file_list(file_list, output_file)
+        # if st.button("Process File", key=line.strip()):
+        #     # Copy Files to Target
+        #     logging.info(f"File {line.strip()} processed successfully.")
+        #     st.success(f"✅ File {line.strip()} processed successfully.")
+
+if st.button("Verify Target Folder"):
+    target_folder = st.text_input("Enter the target folder:", "C:\\Repos\\Data\\Curves\\Target2025M03")
+    if os.path.exists(target_folder):
+        logging.info(f"Target folder {target_folder} exists.")
+        st.info(f"✅ Target folder {target_folder} exists.")
+    else:
+        logging.error(f"Target folder {target_folder} does not exist.")
+        st.error(f"❌ Target folder {target_folder} does not exist.")	
 
 
 if st.button("Copy Files"):
     # Copy files from the plan file to the target folder
-    plan_file = "plan/2025M01_plan.json"
+    
     target_folder = "C:\\Repos\\Data\\Curves\\Target2025M03"
     copy_files_from_plan(plan_file, target_folder)
 
     st.markdown(f"Files are available in the target folder. \n```{target_folder}```")
+
+
